Add second devcontainer
To use the px4 simulator in a container, another ton of dependencies is required, so I made a second devcontainer, inheriting from the first one. Instructions how to use it will be provided in the fcc_bridge reop.
I had to change the folder structure to make this possible, so now vsCode asks you which devcontainer to start.